+++ terraform/modules/lambda/outputs.tf
@@ -10,22 +10,22 @@ output "dlq_url" {
 
 output "function_arn" {
   description = "ARN of the Lambda function."
-  value       = aws_lambda_function.gif_generator_lambda.arn
+  value       = aws_lambda_function.workload.arn
 }
 
 output "function_name" {
   description = "Name of the Lambda function."
-  value       = aws_lambda_function.gif_generator_lambda.function_name
+  value       = aws_lambda_function.workload.function_name
 }
 
 output "image_uri" {
   description = "Image URI currently configured for Lambda."
-  value       = aws_lambda_function.gif_generator_lambda.image_uri
+  value       = aws_lambda_function.workload.image_uri
 }
 
 output "invoke_arn" {
   description = "Invoke ARN used by API Gateway integration."
-  value       = aws_lambda_function.gif_generator_lambda.invoke_arn
+  value       = aws_lambda_function.workload.invoke_arn
 }
 
 output "role_arn" {